# docs/swagger.yaml
basePath: /
definitions:
  model.Subscription:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        description: |-
          @json id
          @format uuid
        example: 4658b3ad-0323-4d4d-854c-05da025bf9ef
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing subscriptions
  title: Subscription Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Returns all subscriptions belonging to a specific user
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
        "400":
          description: Invalid user_id parameter is required
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List all subscriptions for a user
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription with the input payload
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete subscription by ID
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Get subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update subscription fields
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update subscription by ID
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      consumes:
      - application/json
      description: Returns the total sum of subscriptions for a user with optional
        filters
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        type: string
      - description: Service name filter
        in: query
        name: service_name
        type: string
      - description: Start date filter (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date filter (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns total as JSON {\"total\":123}
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Invalid date format
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get total price of subscriptions
      tags:
      - subscriptions
swagger: "2.0"